TwitchAPI/requirements.txt
twitchio==2.8.2
python-dotenv==1.0.0
httpx==0.25.1

requests